refactor: split graph inbox rule classes
